# openapi.json
{"openapi":"3.0.2","info":{"title":"Chat","version":"0.1.0"},"paths":{"/users/":{"get":{"summary":"Read Users","operationId":"read_users_users__get","parameters":[{"required":false,"schema":{"title":"Skip","type":"integer","default":0},"name":"skip","in":"query"},{"required":false,"schema":{"title":"Limit","type":"integer","default":100},"name":"limit","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Read Users Users  Get","type":"array","items":{"$ref":"#/components/schemas/User"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"summary":"Create User","operationId":"create_user_users__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/login":{"post":{"summary":"Login User","operationId":"login_user_users_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/status":{"put":{"summary":"Update User Status","operationId":"update_user_status_users_status_put","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/user/kick":{"get":{"summary":"Kick User","operationId":"kick_user_user_kick_get","parameters":[{"required":true,"schema":{"title":"Receiver Id","type":"integer"},"name":"receiver_id","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/user/ban":{"put":{"summary":"Ban User","operationId":"ban_user_user_ban_put","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserBan"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/status/{status}":{"get":{"summary":"Get Users By Status","operationId":"get_users_by_status_users_status__status__get","parameters":[{"required":true,"schema":{"title":"Status","type":"string"},"name":"status","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/{user_id}":{"get":{"summary":"Read User","operationId":"read_user_users__user_id__get","parameters":[{"required":true,"schema":{"title":"User Id","type":"integer"},"name":"user_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/message/":{"get":{"summary":"Read All Messages","operationId":"read_all_messages_message__get","parameters":[{"required":false,"schema":{"title":"Skip","type":"integer","default":0},"name":"skip","in":"query"},{"required":false,"schema":{"title":"Limit","type":"integer","default":100},"name":"limit","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Read All Messages Message  Get","type":"array","items":{"$ref":"#/components/schemas/Message"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/message/{receiver_id}/":{"post":{"summary":"Create Message From User","operationId":"create_message_from_user_message__receiver_id___post","parameters":[{"required":true,"schema":{"title":"Receiver Id","type":"integer"},"name":"receiver_id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Message"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/message/{receiver_id}/{sender_id}/":{"get":{"summary":"Read Messages To User From","operationId":"read_messages_to_user_from_message__receiver_id___sender_id___get","parameters":[{"required":true,"schema":{"title":"Receiver Id","type":"integer"},"name":"receiver_id","in":"path"},{"required":true,"schema":{"title":"Sender Id","type":"integer"},"name":"sender_id","in":"path"},{"required":false,"schema":{"title":"Skip","type":"integer","default":0},"name":"skip","in":"query"},{"required":false,"schema":{"title":"Limit","type":"integer","default":100},"name":"limit","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Read Messages To User From Message  Receiver Id   Sender Id   Get","type":"array","items":{"$ref":"#/components/schemas/Message"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/message/{receiver_id}/{sender_id}/{from_date}":{"get":{"summary":"Read Messages To User From Date","operationId":"read_messages_to_user_from_date_message__receiver_id___sender_id___from_date__get","parameters":[{"required":true,"schema":{"title":"Receiver Id","type":"integer"},"name":"receiver_id","in":"path"},{"required":true,"schema":{"title":"Sender Id","type":"integer"},"name":"sender_id","in":"path"},{"required":true,"schema":{"title":"From Date","type":"string","format":"date-time"},"name":"from_date","in":"path"},{"required":false,"schema":{"title":"Skip","type":"integer","default":0},"name":"skip","in":"query"},{"required":false,"schema":{"title":"Limit","type":"integer","default":100},"name":"limit","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Read Messages To User From Date Message  Receiver Id   Sender Id   From Date  Get","type":"array","items":{"$ref":"#/components/schemas/Message"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/message/{message_id}":{"delete":{"summary":"Delete Message By Id","operationId":"delete_message_by_id_message__message_id__delete","parameters":[{"required":true,"schema":{"title":"Message Id","type":"integer"},"name":"message_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"Message":{"title":"Message","required":["msg_content","from_usr","id","to_usr","date"],"type":"object","properties":{"msg_content":{"title":"Msg Content","type":"string"},"from_usr":{"title":"From Usr","type":"integer"},"id":{"title":"Id","type":"integer"},"to_usr":{"title":"To Usr","type":"integer"},"date":{"title":"Date","type":"string","format":"date-time"}}},"MessageCreate":{"title":"MessageCreate","required":["msg_content","from_usr"],"type":"object","properties":{"msg_content":{"title":"Msg Content","type":"string"},"from_usr":{"title":"From Usr","type":"integer"}}},"User":{"title":"User","required":["login","id","is_active","is_banned"],"type":"object","properties":{"login":{"title":"Login","type":"string"},"id":{"title":"Id","type":"integer"},"is_active":{"title":"Is Active","type":"boolean"},"is_banned":{"title":"Is Banned","type":"boolean"}}},"UserBan":{"title":"UserBan","required":["id","is_banned"],"type":"object","properties":{"id":{"title":"Id","type":"integer"},"is_banned":{"title":"Is Banned","type":"boolean"}}},"UserCreate":{"title":"UserCreate","required":["login","password"],"type":"object","properties":{"login":{"title":"Login","type":"string"},"password":{"title":"Password","type":"string"}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"type":"string"}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}}}}